requirements.txt
Django==5.1.15
psycopg[binary]==3.3.3
python-dotenv>=1.0.0
Pillow>=10.0.0
djangorestframework>=3.15.0
drf-spectacular>=0.27.0
tzdata>=2025.1

# GDAL / GEOS are installed at OS level.
# On Ubuntu: apt-get install -y gdal-bin libgdal-dev libgeos-dev
# Python binding version must match installed GDAL:
# gdal==<version matching gdal-config --version>
